fix room out

+++ src/chat/routes/api_room.js
@@ -7,9 +7,9 @@ const {room, room_members, room_chats, user, friend} = require('../models');
 let chat = require('../model/chat');
 
 //Create room
-router.post('/', async function ( req, res ) {
-    let { userIds, roomName } = req.body;
-    if ( !userIds ) {
+router.post('/', async function (req, res) {
+    let {userIds, roomName} = req.body;
+    if (!userIds) {
         res.status(200).json({
             message: "방의 멤버를 정해주세요."
         });
@@ -112,16 +112,6 @@ router.get('/:userId/:roomId', async function (req, res, next) {
 
     let roomInfo = await handleDb.readRoom(userID, roomID);
 
-    /**
-     * roomInfo = {
-     *  "userName":  user_in_db.name,
-     *  "userId": userId,
-     *  "roomName": room_members_in_db.room_name,
-     *  "roomId": roomId,
-     *  "chatList": chatList
-     *  }
-     */
-
     try {
         res.status(200).send(roomInfo);
     } catch (err) {
@@ -136,69 +126,36 @@ router.get('/out/:userId/:roomId', async function (req, res, next) {
 
     const userId = req.params.userId;
     const roomId = req.params.roomId;
+
+    room_members.destroy({
+        where: {
+            user_id: userId,
+            room_id: roomId
+        }
+    }).then((result) => {
+        console.log("room_members", result, "행 삭제 완료")
+    }).catch((err) => {
+        console.log("room_members 삭제 실패", err)
+    });
+
     let isP2P = false;
     let user_in_db = await user.findByPk(userId);
 
-    if (user_in_db.myroom === roomId) {
+    if (user_in_db.myroom === roomId) { //개인 채팅방
         user_in_db.myroom = null;
         user_in_db.save();
     } else {
-        let userFriend = await friend.count({
+        let userFriend = await friend.count({ // 1:1 채팅방
             where: {user: userId, room: roomId}
         });
 
         if (userFriend > 0) isP2P = true;
-
     }
 
     room_members.count({
         where: {room_id: roomId}
     }).then(async (c) => {
-        if (c > 1) {
-            room_members.destroy({
-                where: {
-                    user_id: userId,
-                    room_id: roomId
-                }
-            }).then((result) => {
-                console.log("room_members", result, "행 삭제 완료")
-            }).catch((err) => {
-                console.log("room_members 삭제 실패", err)
-            })
-
-        } else {
-            room_members.destroy({
-                where: {
-                    user_id: userId,
-                    room_id: roomId
-                }
-            }).then((result) => {
-                console.log("room_members", result, "행 삭제 완료")
-            }).catch((err) => {
-                console.log("room_members 삭제 실패", err)
-            });
-
-
-            room_chats.destroy({
-                where: {
-                    room_id: roomId
-                }
-            }).then((result) => {
-                console.log("room_chats", result, "행 삭제 완료")
-            }).catch((err) => {
-                console.log("room_chats 디비 삭제 실패", err)
-            });
-
-            chat.deleteMany({room: roomId})
-                .then((result) => {
-                    console.log("chat 디비 모두 삭제 성공", result)
-                })
-                .catch((err) => {
-                    console.log("chat 디비 모두 삭제 실패", err)
-                });
-
-            if (isP2P) return;
-
+        if (!isP2P && c <= 1) {
             room.destroy({
                 where: {
                     id: roomId
@@ -208,7 +165,6 @@ router.get('/out/:userId/:roomId', async function (req, res, next) {
             }).catch((err) => {
                 console.log("room 디비 삭제 실패", err)
             });
-
         }
     });
     res.status(200).send();
